+++ cmd/service_deploy.go
@@ -179,7 +179,10 @@ func deployImage(operation *ServiceDeployOperation) string {
 
 func getDockerServiceFromComposeFile(dockerComposeFile string) *dockercompose.Service {
 	//read the compose file configuration
-	composeFile := dockercompose.Read(dockerComposeFile)
+	composeFile, err := dockercompose.Read(dockerComposeFile)
+	if err != nil {
+		console.ErrorExit(err, "error reading docker compose file")
+	}
 
 	//determine which docker-compose service/container to deploy
 	_, dockerService := getDockerServiceToDeploy(&composeFile.Data)